+++ src/services/psql.ts
@@ -1,0 +1,32 @@
+import * as pg from 'pg';
+
+const Client = pg.Client;
+
+async function getBinId(urlPath: string) {
+  const client = new Client({
+    "database": "rhh",
+  })
+  await client.connect()
+
+  const result = await client.query("SELECT * FROM bins WHERE url_path = $1", [urlPath])
+
+  await client.end()
+
+  if (result.rowCount != 1) {
+    return undefined
+  } else {
+    return result.rows[0].id
+  }
+
+}
+
+async function saveRequest(mongoId: string, binId: number, http_method: string, http_path: string) {
+  const client = new Client({
+    "database": "rhh",
+  })
+  await client.connect()
+
+  client.query("INSERT INTO requests (mongo_id, bin_id, http_method, http_path) VALUES ($1, $2, $3, $4)", [mongoId, binId, http_method, http_path])
+}
+
+export { getBinId, saveRequest }
